added modification for corrections officer

+++ result_system/serializers.py
@@ -38,10 +38,8 @@ class AssessmentSerializer(serializers.ModelSerializer):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        # Get the result from context (passed by the view)
         result = self.context.get('result')
         if result:
-            # Get students enrolled in the course linked to the result
             enrolled_students = Student.objects.filter(
                 enrolled_student__course=result.course
             ).distinct()
@@ -53,29 +51,31 @@ class AssessmentSerializer(serializers.ModelSerializer):
 
 
 class SubmitResultSerializer(serializers.Serializer):
-    result_id = serializers.IntegerField()
+    result_id = serializers.IntegerField(read_only=True)
 
     def save(self, **kwargs):
         with transaction.atomic():
-            result_id = self.validated_data['result_id']
-            lecturer_id = User.objects.get(id=self.context['lecturer_id'])
-            submitted_result = SubmittedResult.objects.create(
-                lecturer=lecturer_id)
-            assessments = Assessment.objects.filter(result_id=result_id)
-            submitted_result_score = [
-                SubmittedResultScore(
-                    submitted_result=submitted_result,
-                    student=assessment.student,
-                    ca_slot1=assessment.ca_slot1,
-                    ca_slot2=assessment.ca_slot2,
-                    ca_slot3=assessment.ca_slot3,
-                    ca_slot4=assessment.ca_slot4,
-                    exam_mark=assessment.exam_mark
-                ) for assessment in assessments
-            ]
+            result_id = self.context['result_id']
+            if Result.objects.filter(id=result_id).exists():
+                lecturer_id = User.objects.get(id=self.context['lecturer_id'])
+                submitted_result = SubmittedResult.objects.create(
+                    lecturer=lecturer_id)
+                assessments = Assessment.objects.filter(result_id=result_id)
+                submitted_result_score = [
+                    SubmittedResultScore(
+                        submitted_result=submitted_result,
+                        student=assessment.student,
+                        ca_slot1=assessment.ca_slot1,
+                        ca_slot2=assessment.ca_slot2,
+                        ca_slot3=assessment.ca_slot3,
+                        ca_slot4=assessment.ca_slot4,
+                        exam_mark=assessment.exam_mark
+                    ) for assessment in assessments
+                ]
 
-            SubmittedResultScore.objects.bulk_create(submitted_result_score)
-            Result.objects.filter(id=result_id).delete()
+                SubmittedResultScore.objects.bulk_create(
+                    submitted_result_score)
+                Result.objects.filter(id=result_id).delete()
 
 
 class SubmittedResultSerializer(serializers.ModelSerializer):
@@ -93,9 +93,6 @@ class SubmittedResultScoreSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
         result_id = self.context['submitted_result_id']
         return Assessment.objects.create(result_id=result_id, **validated_data)
-
-
-
 
 
 class UserSerializer(serializers.ModelSerializer):
